tool/requirements.txt
google-cloud-firestore>=2.1.0
